Fix binary Release file generation
